+++ .github/workflows/test-package-build.yml
@@ -51,8 +51,8 @@ jobs:
       - name: Test installation from tarball
         run: |
           pip install ${{ steps.build-dist.outputs.sdist-path }}
-          python -c "import unitycatalog.ai; print(unitycatalog.ai.__version__)"
+          python -c "import ucai.core; print(ucai.core.__version__)"
       - name: Test installation from wheel
         run: |
           pip install --force-reinstall ${{ steps.build-dist.outputs.wheel-path }}
-          python -c "import unitycatalog.ai; print(unitycatalog.ai.__version__)"
+          python -c "import ucai.core; print(ucai.core.__version__)"
