+++ integrations/langchain/pyproject.toml
@@ -11,14 +11,14 @@ requires-python = ">=3.9"
 dependencies = [
     "langchain>=0.2.0",
     "langchain-community>=0.2.0",
-    "unitycatalog-ai",
+    "ucai-core",
 ]
 
 [project.optional-dependencies]
 dev = [
   "pytest",
   "typing_extensions",
-  # temp needed before unitycatalog-ai[databricks] is released
+  # temp needed before ucai-core[databricks] is released
   "databricks-sdk>=0.32.0",
   "databricks-connect",
   "ruff==0.6.4"
