+++ integrations/llama_index/src/ucai_llamaindex/toolkit.py
@@ -4,9 +4,9 @@ from typing import Any, Callable, Dict, List, Optional
 from llama_index.core.tools import FunctionTool
 from llama_index.core.tools.types import ToolMetadata
 from pydantic import BaseModel, ConfigDict, Field, model_validator
-from unitycatalog.ai.client import BaseFunctionClient
-from unitycatalog.ai.utils.client_utils import validate_or_set_default_client
-from unitycatalog.ai.utils.function_processing_utils import (
+from ucai.core.client import BaseFunctionClient
+from ucai.core.utils.client_utils import validate_or_set_default_client
+from ucai.core.utils.function_processing_utils import (
     generate_function_input_params_schema,
     get_tool_name,
     process_function_names,
