+++ integrations/openai/pyproject.toml
@@ -10,7 +10,7 @@ license = { text="Apache-2.0" }
 requires-python = ">=3.9"
 dependencies = [
     "openai>=1.46.1",
-    "unitycatalog-ai",
+    "ucai-core",
     "pydantic<3,>=2",
 ]
 
@@ -18,7 +18,7 @@ dependencies = [
 dev = [
   "pytest",
   "typing_extensions",
-  # temp needed before unitycatalog-ai[databricks] is released
+  # temp needed before ucai-core[databricks] is released
   "databricks-sdk>=0.32.0",
   "databricks-connect",
   "ruff==0.6.4"
