+++ integrations/openai/src/ucai_openai/toolkit.py
@@ -3,9 +3,9 @@ from typing import Any, Dict, List, Optional
 from openai import pydantic_function_tool
 from openai.types.chat import ChatCompletionToolParam
 from pydantic import BaseModel, ConfigDict, Field, model_validator
-from unitycatalog.ai.client import BaseFunctionClient
-from unitycatalog.ai.utils.client_utils import validate_or_set_default_client
-from unitycatalog.ai.utils.function_processing_utils import (
+from ucai.core.client import BaseFunctionClient
+from ucai.core.utils.client_utils import validate_or_set_default_client
+from ucai.core.utils.function_processing_utils import (
     generate_function_input_params_schema,
     get_tool_name,
     process_function_names,
