+++ src/ucai/core/__init__.py
@@ -1,0 +1,3 @@
+from ucai.core.version import VERSION
+
+__version__ = VERSION
