+++ src/ucai/core/client.py
@@ -6,7 +6,7 @@ from copy import deepcopy
 from dataclasses import dataclass
 from typing import Any, Dict, Literal, Optional
 
-from unitycatalog.ai.paged_list import PagedList
+from ucai.core.paged_list import PagedList
 
 _logger = logging.getLogger(__name__)
