+++ src/ucai/test_utils/client_utils.py
@@ -4,8 +4,8 @@ from unittest import mock
 
 import pytest
 
-from unitycatalog.ai.client import set_uc_function_client
-from unitycatalog.ai.databricks import DatabricksFunctionClient
+from ucai.core.client import set_uc_function_client
+from ucai.core.databricks import DatabricksFunctionClient
 
 USE_SERVERLESS = "USE_SERVERLESS"
 
@@ -23,7 +23,7 @@ def requires_databricks(test_func):
 @pytest.fixture
 def client() -> DatabricksFunctionClient:
     with mock.patch(
-        "unitycatalog.ai.databricks.get_default_databricks_workspace_client",
+        "ucai.core.databricks.get_default_databricks_workspace_client",
         return_value=mock.Mock(),
     ):
         return DatabricksFunctionClient(warehouse_id="warehouse_id", cluster_id="cluster_id")
@@ -36,7 +36,7 @@ def serverless_client() -> DatabricksFunctionClient:
 
 def get_client() -> DatabricksFunctionClient:
     with mock.patch(
-        "unitycatalog.ai.databricks.get_default_databricks_workspace_client",
+        "ucai.core.databricks.get_default_databricks_workspace_client",
         return_value=mock.Mock(),
     ):
         if os.environ.get(USE_SERVERLESS, "false").lower() == "true":
