--- src/unitycatalog/ai/__init__.py
@@ -1,3 +1,0 @@
-from unitycatalog.ai.version import VERSION
-
-__version__ = VERSION
